+++ Assignment-1.08/io.cpp
@@ -13,7 +13,8 @@
 /* Same ugly hack we did in path.c */
 static dungeon *thedungeon;
 
-typedef struct io_message {
+typedef struct io_message
+{
   /* Will print " --more-- " at end of line when another message follows. *
    * Leave 10 extra spaces for that.                                      */
   char msg[71];
@@ -43,7 +44,8 @@ void io_reset_terminal(void)
 {
   endwin();
 
-  while (io_head) {
+  while (io_head)
+  {
     io_tail = io_head;
     io_head = io_head->next;
     free(io_tail);
@@ -56,7 +58,8 @@ void io_queue_message(const char *format, ...)
   io_message_t *tmp;
   va_list ap;
 
-  if (!(tmp = (io_message_t *) malloc(sizeof (*tmp)))) {
+  if (!(tmp = (io_message_t *)malloc(sizeof(*tmp))))
+  {
     perror("malloc");
     exit(1);
   }
@@ -65,13 +68,16 @@ void io_queue_message(const char *format, ...)
 
   va_start(ap, format);
 
-  vsnprintf(tmp->msg, sizeof (tmp->msg), format, ap);
+  vsnprintf(tmp->msg, sizeof(tmp->msg), format, ap);
 
   va_end(ap);
 
-  if (!io_head) {
+  if (!io_head)
+  {
     io_head = io_tail = tmp;
-  } else {
+  }
+  else
+  {
     io_tail->next = tmp;
     io_tail = tmp;
   }
@@ -79,13 +85,15 @@ void io_queue_message(const char *format, ...)
 
 static void io_print_message_queue(uint32_t y, uint32_t x)
 {
-  while (io_head) {
+  while (io_head)
+  {
     io_tail = io_head;
     attron(COLOR_PAIR(COLOR_CYAN));
     mvprintw(y, x, "%-80s", io_head->msg);
     attroff(COLOR_PAIR(COLOR_CYAN));
     io_head = io_head->next;
-    if (io_head) {
+    if (io_head)
+    {
       attron(COLOR_PAIR(COLOR_CYAN));
       mvprintw(y, x + 70, "%10s", " --more-- ");
       attroff(COLOR_PAIR(COLOR_CYAN));
@@ -101,13 +109,20 @@ void io_display_tunnel(dungeon *d)
 {
   uint32_t y, x;
   clear();
-  for (y = 0; y < DUNGEON_Y; y++) {
-    for (x = 0; x < DUNGEON_X; x++) {
-      if (charxy(x, y) == d->PC) {
+  for (y = 0; y < DUNGEON_Y; y++)
+  {
+    for (x = 0; x < DUNGEON_X; x++)
+    {
+      if (charxy(x, y) == d->PC)
+      {
         mvaddch(y + 1, x, charxy(x, y)->symbol);
-      } else if (hardnessxy(x, y) == 255) {
+      }
+      else if (hardnessxy(x, y) == 255)
+      {
         mvaddch(y + 1, x, '*');
-      } else {
+      }
+      else
+      {
         mvaddch(y + 1, x, '0' + (d->pc_tunnel[y][x] % 10));
       }
     }
@@ -119,13 +134,20 @@ void io_display_distance(dungeon *d)
 {
   uint32_t y, x;
   clear();
-  for (y = 0; y < DUNGEON_Y; y++) {
-    for (x = 0; x < DUNGEON_X; x++) {
-      if (charxy(x, y)) {
+  for (y = 0; y < DUNGEON_Y; y++)
+  {
+    for (x = 0; x < DUNGEON_X; x++)
+    {
+      if (charxy(x, y))
+      {
         mvaddch(y + 1, x, charxy(x, y)->symbol);
-      } else if (hardnessxy(x, y) != 0) {
+      }
+      else if (hardnessxy(x, y) != 0)
+      {
         mvaddch(y + 1, x, ' ');
-      } else {
+      }
+      else
+      {
         mvaddch(y + 1, x, '0' + (d->pc_distance[y][x] % 10));
       }
     }
@@ -134,23 +156,23 @@ void io_display_distance(dungeon *d)
 }
 
 static char hardness_to_char[] =
-  "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
+    "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
 void io_display_hardness(dungeon *d)
 {
   uint32_t y, x;
   clear();
-  for (y = 0; y < DUNGEON_Y; y++) {
-    for (x = 0; x < DUNGEON_X; x++) {
+  for (y = 0; y < DUNGEON_Y; y++)
+  {
+    for (x = 0; x < DUNGEON_X; x++)
+    {
       /* Maximum hardness is 255.  We have 62 values to display it, but *
        * we only want one zero value, so we need to cover [1,255] with  *
        * 61 values, which gives us a divisor of 254 / 61 = 4.164.       *
        * Generally, we want to avoid floating point math, but this is   *
        * not gameplay, so we'll make an exception here to get maximal   *
        * hardness display resolution.                                   */
-      mvaddch(y + 1, x, (d->hardness[y][x]                             ?
-                         hardness_to_char[1 + (int) ((d->hardness[y][x] /
-                                                      4.2))] : ' '));
+      mvaddch(y + 1, x, (d->hardness[y][x] ? hardness_to_char[1 + (int)((d->hardness[y][x] / 4.2))] : ' '));
     }
   }
   refresh();
@@ -158,8 +180,8 @@ void io_display_hardness(dungeon *d)
 
 static int compare_monster_distance(const void *v1, const void *v2)
 {
-  const character *const *c1 = (const character *const *) v1;
-  const character *const *c2 = (const character *const *) v2;
+  const character *const *c1 = (const character *const *)v1;
+  const character *const *c2 = (const character *const *)v2;
 
   return (thedungeon->pc_distance[(*c1)->position[dim_y]]
                                  [(*c1)->position[dim_x]] -
@@ -172,12 +194,15 @@ static character *io_nearest_visible_monster(dungeon *d)
   character **c, *n;
   uint32_t x, y, count, i;
 
-  c = (character **) malloc(d->num_monsters * sizeof (*c));
+  c = (character **)malloc(d->num_monsters * sizeof(*c));
 
   /* Get a linear list of monsters */
-  for (count = 0, y = 1; y < DUNGEON_Y - 1; y++) {
-    for (x = 1; x < DUNGEON_X - 1; x++) {
-      if (d->character_map[y][x] && d->character_map[y][x] != d->PC) {
+  for (count = 0, y = 1; y < DUNGEON_Y - 1; y++)
+  {
+    for (x = 1; x < DUNGEON_X - 1; x++)
+    {
+      if (d->character_map[y][x] && d->character_map[y][x] != d->PC)
+      {
         c[count++] = d->character_map[y][x];
       }
     }
@@ -185,10 +210,12 @@ static character *io_nearest_visible_monster(dungeon *d)
 
   /* Sort it by distance from PC */
   thedungeon = d;
-  qsort(c, count, sizeof (*c), compare_monster_distance);
+  qsort(c, count, sizeof(*c), compare_monster_distance);
 
-  for (n = NULL, i = 0; i < count; i++) {
-    if (can_see(d, character_get_pos(d->PC), character_get_pos(c[i]), 1, 0)) {
+  for (n = NULL, i = 0; i < count; i++)
+  {
+    if (can_see(d, character_get_pos(d->PC), character_get_pos(c[i]), 1, 0))
+    {
       n = c[i];
       break;
     }
@@ -207,30 +234,130 @@ void io_display(dungeon *d)
   int32_t visible_monsters;
 
   clear();
-  for (visible_monsters = -1, y = 0; y < 21; y++) {
-    for (x = 0; x < 80; x++) {
-      if ((illuminated = is_illuminated(d->PC, y, x))) {
+  for (visible_monsters = -1, y = 0; y < 21; y++)
+  {
+    for (x = 0; x < 80; x++)
+    {
+      if ((illuminated = is_illuminated(d->PC, y, x)))
+      {
         attron(A_BOLD);
       }
       if (d->character_map[y][x] &&
-           can_see(d,
+          can_see(d,
                   character_get_pos(d->PC),
                   character_get_pos(d->character_map[y][x]),
-                  1, 0)) {
-              
-              //Lee's
-              if( y != (unsigned)d->PC->position[dim_y] || x != (unsigned)d->PC->position[dim_x]){
-              attron(COLOR_PAIR(static_cast<npc*>(d->character_map[y][x])->color[0]));
-              mvaddch(y + 1, x,character_get_symbol(d->character_map[y][x]));
-              attroff(COLOR_PAIR(static_cast<npc*>(d->character_map[y][x])->color[0]));
-              }
-              else{
-                mvaddch(y + 1, x,character_get_symbol(d->character_map[y][x]));
-              }
+                  1, 0))
+      {
+
+        //Lee's
+        if (y != (unsigned)d->PC->position[dim_y] || x != (unsigned)d->PC->position[dim_x])
+        {
+          attron(COLOR_PAIR(static_cast<npc *>(d->character_map[y][x])->color[0]));
+          mvaddch(y + 1, x, character_get_symbol(d->character_map[y][x]));
+          attroff(COLOR_PAIR(static_cast<npc *>(d->character_map[y][x])->color[0]));
+        }
+        else
+        {
+          mvaddch(y + 1, x, character_get_symbol(d->character_map[y][x]));
+        }
 
         visible_monsters++;
-      } else {
-        switch (pc_learned_terrain(d->PC, y, x)) {
+      }
+
+      //Leyuan
+      else if (d->object_map[y][x] && can_see(d, character_get_pos(d->PC), d->object_map[y][x]->position, 1, 0))
+      {
+        //get the symbol
+        char obj_symbol = ' ';
+        object_type_t obj_type = (d->object_map[y][x])->type;
+        if (obj_type == objtype_no_type)
+        {
+          obj_symbol = '*';
+        }
+        else if (obj_type == objtype_WEAPON)
+        {
+          obj_symbol = '|';
+        }
+        else if (obj_type == objtype_OFFHAND)
+        {
+          obj_symbol = ')';
+        }
+        else if (obj_type == objtype_RANGED)
+        {
+          obj_symbol = '}';
+        }
+        else if (obj_type == objtype_LIGHT)
+        {
+          obj_symbol = '_';
+        }
+        else if (obj_type == objtype_ARMOR)
+        {
+          obj_symbol = '[';
+        }
+        else if (obj_type == objtype_HELMET)
+        {
+          obj_symbol = ']';
+        }
+        else if (obj_type == objtype_CLOAK)
+        {
+          obj_symbol = '(';
+        }
+        else if (obj_type == objtype_GLOVES)
+        {
+          obj_symbol = '{';
+        }
+        else if (obj_type == objtype_BOOTS)
+        {
+          obj_symbol = '\\';
+        }
+        else if (obj_type == objtype_AMULET)
+        {
+          obj_symbol = '\"';
+        }
+        else if (obj_type == objtype_RING)
+        {
+          obj_symbol = '=';
+        }
+        else if (obj_type == objtype_SCROLL)
+        {
+          obj_symbol = '~';
+        }
+        else if (obj_type == objtype_BOOK)
+        {
+          obj_symbol = '?';
+        }
+        else if (obj_type == objtype_FLASK)
+        {
+          obj_symbol = '!';
+        }
+        else if (obj_type == objtype_GOLD)
+        {
+          obj_symbol = '$';
+        }
+        else if (obj_type == objtype_AMMUNITION)
+        {
+          obj_symbol = '/';
+        }
+        else if (obj_type == objtype_FOOD)
+        {
+          obj_symbol = ',';
+        }
+        else if (obj_type == objtype_WAND)
+        {
+          obj_symbol = '-';
+        }
+        else if (obj_type == objtype_CONTAINER)
+        {
+          obj_symbol = '%';
+        }
+        attron(COLOR_PAIR(static_cast<real_object *>(d->object_map[y][x])->color));
+        mvaddch(y + 1, x, obj_symbol);
+        attroff(COLOR_PAIR(static_cast<real_object *>(d->object_map[y][x])->color));
+      }
+      else
+      {
+        switch (pc_learned_terrain(d->PC, y, x))
+        {
         case ter_wall:
         case ter_wall_immutable:
         case ter_unknown:
@@ -253,12 +380,13 @@ void io_display(dungeon *d)
           mvaddch(y + 1, x, '>');
           break;
         default:
- /* Use zero as an error symbol, since it stands out somewhat, and it's *
+          /* Use zero as an error symbol, since it stands out somewhat, and it's *
   * not otherwise used.                                                 */
           mvaddch(y + 1, x, '0');
         }
       }
-      if (illuminated) {
+      if (illuminated)
+      {
         attroff(A_BOLD);
       }
     }
@@ -269,25 +397,26 @@ void io_display(dungeon *d)
   mvprintw(22, 1, "%d known %s.", visible_monsters,
            visible_monsters > 1 ? "monsters" : "monster");
   mvprintw(22, 30, "Nearest visible monster: ");
-  if ((c = io_nearest_visible_monster(d))) {
+  if ((c = io_nearest_visible_monster(d)))
+  {
     attron(COLOR_PAIR(COLOR_RED));
     mvprintw(22, 55, "%c at %d %c by %d %c.",
              c->symbol,
              abs(c->position[dim_y] - d->PC->position[dim_y]),
-             ((c->position[dim_y] - d->PC->position[dim_y]) <= 0 ?
-              'N' : 'S'),
+             ((c->position[dim_y] - d->PC->position[dim_y]) <= 0 ? 'N' : 'S'),
              abs(c->position[dim_x] - d->PC->position[dim_x]),
-             ((c->position[dim_x] - d->PC->position[dim_x]) <= 0 ?
-              'W' : 'E'));
+             ((c->position[dim_x] - d->PC->position[dim_x]) <= 0 ? 'W' : 'E'));
     attroff(COLOR_PAIR(COLOR_RED));
-  } else {
+  }
+  else
+  {
     attron(COLOR_PAIR(COLOR_BLUE));
     mvprintw(22, 55, "NONE.");
     attroff(COLOR_PAIR(COLOR_BLUE));
   }
-  
+
   io_print_message_queue(0, 0);
-  
+
   refresh();
 }
 
@@ -297,20 +426,119 @@ void io_display_no_fog(dungeon *d)
   character *c;
 
   clear();
-  for (y = 0; y < 21; y++) {
-    for (x = 0; x < 80; x++) {
-      if (d->character_map[y][x]) {
+  for (y = 0; y < 21; y++)
+  {
+    for (x = 0; x < 80; x++)
+    {
+      if (d->character_map[y][x])
+      {
         //Lee's
-        if( y != (unsigned)d->PC->position[dim_y] || x != (unsigned)d->PC->position[dim_x]){
-              attron(COLOR_PAIR(static_cast<npc*>(d->character_map[y][x])->color[0]));
-              mvaddch(y + 1, x,character_get_symbol(d->character_map[y][x]));
-              attroff(COLOR_PAIR(static_cast<npc*>(d->character_map[y][x])->color[0]));
-              }
-              else{
-                mvaddch(y + 1, x,character_get_symbol(d->character_map[y][x]));
-              }
-      } else {
-        switch (mapxy(x, y)) {
+        if (y != (unsigned)d->PC->position[dim_y] || x != (unsigned)d->PC->position[dim_x])
+        {
+          attron(COLOR_PAIR(static_cast<npc *>(d->character_map[y][x])->color[0]));
+          mvaddch(y + 1, x, character_get_symbol(d->character_map[y][x]));
+          attroff(COLOR_PAIR(static_cast<npc *>(d->character_map[y][x])->color[0]));
+        }
+        else
+        {
+          mvaddch(y + 1, x, character_get_symbol(d->character_map[y][x]));
+        }
+      }
+
+      //Leyuan
+      else if (d->object_map[y][x])
+      {
+        //get the symbol
+        char obj_symbol = ' ';
+        object_type_t obj_type = (d->object_map[y][x])->type;
+        if (obj_type == objtype_no_type)
+        {
+          obj_symbol = '*';
+        }
+        else if (obj_type == objtype_WEAPON)
+        {
+          obj_symbol = '|';
+        }
+        else if (obj_type == objtype_OFFHAND)
+        {
+          obj_symbol = ')';
+        }
+        else if (obj_type == objtype_RANGED)
+        {
+          obj_symbol = '}';
+        }
+        else if (obj_type == objtype_LIGHT)
+        {
+          obj_symbol = '_';
+        }
+        else if (obj_type == objtype_ARMOR)
+        {
+          obj_symbol = '[';
+        }
+        else if (obj_type == objtype_HELMET)
+        {
+          obj_symbol = ']';
+        }
+        else if (obj_type == objtype_CLOAK)
+        {
+          obj_symbol = '(';
+        }
+        else if (obj_type == objtype_GLOVES)
+        {
+          obj_symbol = '{';
+        }
+        else if (obj_type == objtype_BOOTS)
+        {
+          obj_symbol = '\\';
+        }
+        else if (obj_type == objtype_AMULET)
+        {
+          obj_symbol = '\"';
+        }
+        else if (obj_type == objtype_RING)
+        {
+          obj_symbol = '=';
+        }
+        else if (obj_type == objtype_SCROLL)
+        {
+          obj_symbol = '~';
+        }
+        else if (obj_type == objtype_BOOK)
+        {
+          obj_symbol = '?';
+        }
+        else if (obj_type == objtype_FLASK)
+        {
+          obj_symbol = '!';
+        }
+        else if (obj_type == objtype_GOLD)
+        {
+          obj_symbol = '$';
+        }
+        else if (obj_type == objtype_AMMUNITION)
+        {
+          obj_symbol = '/';
+        }
+        else if (obj_type == objtype_FOOD)
+        {
+          obj_symbol = ',';
+        }
+        else if (obj_type == objtype_WAND)
+        {
+          obj_symbol = '-';
+        }
+        else if (obj_type == objtype_CONTAINER)
+        {
+          obj_symbol = '%';
+        }
+        attron(COLOR_PAIR(static_cast<real_object *>(d->object_map[y][x])->color));
+        mvaddch(y + 1, x, obj_symbol);
+        attroff(COLOR_PAIR(static_cast<real_object *>(d->object_map[y][x])->color));
+      }
+      else
+      {
+        switch (mapxy(x, y))
+        {
         case ter_wall:
         case ter_wall_immutable:
           mvaddch(y + 1, x, ' ');
@@ -332,7 +560,7 @@ void io_display_no_fog(dungeon *d)
           mvaddch(y + 1, x, '>');
           break;
         default:
- /* Use zero as an error symbol, since it stands out somewhat, and it's *
+          /* Use zero as an error symbol, since it stands out somewhat, and it's *
   * not otherwise used.                                                 */
           mvaddch(y + 1, x, '0');
         }
@@ -345,23 +573,24 @@ void io_display_no_fog(dungeon *d)
   mvprintw(22, 1, "%d %s.", d->num_monsters,
            d->num_monsters > 1 ? "monsters" : "monster");
   mvprintw(22, 30, "Nearest visible monster: ");
-  if ((c = io_nearest_visible_monster(d))) {
+  if ((c = io_nearest_visible_monster(d)))
+  {
     attron(COLOR_PAIR(COLOR_RED));
     mvprintw(22, 55, "%c at %d %c by %d %c.",
              c->symbol,
              abs(c->position[dim_y] - d->PC->position[dim_y]),
-             ((c->position[dim_y] - d->PC->position[dim_y]) <= 0 ?
-              'N' : 'S'),
+             ((c->position[dim_y] - d->PC->position[dim_y]) <= 0 ? 'N' : 'S'),
              abs(c->position[dim_x] - d->PC->position[dim_x]),
-             ((c->position[dim_x] - d->PC->position[dim_x]) <= 0 ?
-              'W' : 'E'));
+             ((c->position[dim_x] - d->PC->position[dim_x]) <= 0 ? 'W' : 'E'));
     attroff(COLOR_PAIR(COLOR_RED));
-  } else {
+  }
+  else
+  {
     attron(COLOR_PAIR(COLOR_BLUE));
     mvprintw(22, 55, "NONE.");
     attroff(COLOR_PAIR(COLOR_BLUE));
   }
-  
+
   io_print_message_queue(0, 0);
 }
 
@@ -392,11 +621,115 @@ uint32_t io_teleport_pc(dungeon *d)
   mvaddch(dest[dim_y] + 1, dest[dim_x], '*');
   refresh();
 
-  while ((c = getch()) != 'g' && c != '.' && c != 'r') {
-    if (charpair(dest)) {
-      actual = character_get_symbol(charpair(dest));
-    } else {
-      switch (mappair(dest)) {
+  while ((c = getch()) != 'g' && c != '.' && c != 'r')
+  {
+    if (charpair(dest))
+    {
+      //Leyuan*
+      if (dest[dim_y] != (int16_t)d->PC->position[dim_y] || dest[dim_x] != (int16_t)d->PC->position[dim_x])
+      {
+        attron(COLOR_PAIR(static_cast<npc *>(d->character_map[dest[dim_y]][dest[dim_x]])->color[0]));
+        mvaddch(dest[dim_y] + 1, dest[dim_x], character_get_symbol(d->character_map[dest[dim_y]][dest[dim_x]]));
+        attroff(COLOR_PAIR(static_cast<npc *>(d->character_map[dest[dim_y]][dest[dim_x]])->color[0]));
+      }
+      else
+      {
+        mvaddch(dest[dim_y] + 1, dest[dim_x], character_get_symbol(d->character_map[dest[dim_y]][dest[dim_x]]));
+      }
+    }
+    else if (d->object_map[dest[dim_y]][dest[dim_x]])
+    {
+      //get the symbol
+      char obj_symbol = ' ';
+      object_type_t obj_type = (d->object_map[dest[dim_y]][dest[dim_x]])->type;
+      if (obj_type == objtype_no_type)
+      {
+        obj_symbol = '*';
+      }
+      else if (obj_type == objtype_WEAPON)
+      {
+        obj_symbol = '|';
+      }
+      else if (obj_type == objtype_OFFHAND)
+      {
+        obj_symbol = ')';
+      }
+      else if (obj_type == objtype_RANGED)
+      {
+        obj_symbol = '}';
+      }
+      else if (obj_type == objtype_LIGHT)
+      {
+        obj_symbol = '_';
+      }
+      else if (obj_type == objtype_ARMOR)
+      {
+        obj_symbol = '[';
+      }
+      else if (obj_type == objtype_HELMET)
+      {
+        obj_symbol = ']';
+      }
+      else if (obj_type == objtype_CLOAK)
+      {
+        obj_symbol = '(';
+      }
+      else if (obj_type == objtype_GLOVES)
+      {
+        obj_symbol = '{';
+      }
+      else if (obj_type == objtype_BOOTS)
+      {
+        obj_symbol = '\\';
+      }
+      else if (obj_type == objtype_AMULET)
+      {
+        obj_symbol = '\"';
+      }
+      else if (obj_type == objtype_RING)
+      {
+        obj_symbol = '=';
+      }
+      else if (obj_type == objtype_SCROLL)
+      {
+        obj_symbol = '~';
+      }
+      else if (obj_type == objtype_BOOK)
+      {
+        obj_symbol = '?';
+      }
+      else if (obj_type == objtype_FLASK)
+      {
+        obj_symbol = '!';
+      }
+      else if (obj_type == objtype_GOLD)
+      {
+        obj_symbol = '$';
+      }
+      else if (obj_type == objtype_AMMUNITION)
+      {
+        obj_symbol = '/';
+      }
+      else if (obj_type == objtype_FOOD)
+      {
+        obj_symbol = ',';
+      }
+      else if (obj_type == objtype_WAND)
+      {
+        obj_symbol = '-';
+      }
+      else if (obj_type == objtype_CONTAINER)
+      {
+        obj_symbol = '%';
+      }
+      attron(COLOR_PAIR(static_cast<real_object *>(d->object_map[dest[dim_y]][dest[dim_x]])->color));
+      mvaddch(dest[dim_y] + 1, dest[dim_x], obj_symbol);
+      attroff(COLOR_PAIR(static_cast<real_object *>(d->object_map[dest[dim_y]][dest[dim_x]])->color));
+    }
+    else
+    {
+      switch (mappair(dest))
+      {
       case ter_wall:
       case ter_wall_immutable:
         actual = ' ';
@@ -419,77 +752,89 @@ uint32_t io_teleport_pc(dungeon *d)
         break;
       default:
         break;
-      }      
+      }
+      mvaddch(dest[dim_y] + 1, dest[dim_x], actual);
     }
 
-    mvaddch(dest[dim_y] + 1, dest[dim_x], actual);
-
-    switch (c) {
+    switch (c)
+    {
     case '7':
     case 'y':
     case KEY_HOME:
-      if (dest[dim_y] > 1) {
+      if (dest[dim_y] > 1)
+      {
         dest[dim_y]--;
       }
-      if (dest[dim_x] > 1) {
+      if (dest[dim_x] > 1)
+      {
         dest[dim_x]--;
       }
       break;
     case '8':
     case 'k':
     case KEY_UP:
-      if (dest[dim_y] > 1) {
+      if (dest[dim_y] > 1)
+      {
         dest[dim_y]--;
       }
       break;
     case '9':
     case 'u':
     case KEY_PPAGE:
-      if (dest[dim_y] > 1) {
+      if (dest[dim_y] > 1)
+      {
         dest[dim_y]--;
       }
-      if (dest[dim_x] < DUNGEON_X - 2) {
+      if (dest[dim_x] < DUNGEON_X - 2)
+      {
         dest[dim_x]++;
       }
       break;
     case '6':
     case 'l':
     case KEY_RIGHT:
-      if (dest[dim_x] < DUNGEON_X - 2) {
+      if (dest[dim_x] < DUNGEON_X - 2)
+      {
         dest[dim_x]++;
       }
       break;
     case '3':
     case 'n':
     case KEY_NPAGE:
-      if (dest[dim_y] < DUNGEON_Y - 2) {
+      if (dest[dim_y] < DUNGEON_Y - 2)
+      {
         dest[dim_y]++;
       }
-      if (dest[dim_x] < DUNGEON_X - 2) {
+      if (dest[dim_x] < DUNGEON_X - 2)
+      {
         dest[dim_x]++;
       }
       break;
     case '2':
     case 'j':
     case KEY_DOWN:
-      if (dest[dim_y] < DUNGEON_Y - 2) {
+      if (dest[dim_y] < DUNGEON_Y - 2)
+      {
         dest[dim_y]++;
       }
       break;
     case '1':
     case 'b':
     case KEY_END:
-      if (dest[dim_y] < DUNGEON_Y - 2) {
+      if (dest[dim_y] < DUNGEON_Y - 2)
+      {
         dest[dim_y]++;
       }
-      if (dest[dim_x] > 1) {
+      if (dest[dim_x] > 1)
+      {
         dest[dim_x]--;
       }
       break;
     case '4':
     case 'h':
     case KEY_LEFT:
-      if (dest[dim_x] > 1) {
+      if (dest[dim_x] > 1)
+      {
         dest[dim_x]--;
       }
       break;
@@ -499,16 +844,21 @@ uint32_t io_teleport_pc(dungeon *d)
     refresh();
   }
 
-  if (c == 'r') {
-    do {
+  if (c == 'r')
+  {
+    do
+    {
       dest[dim_x] = rand_range(1, DUNGEON_X - 2);
       dest[dim_y] = rand_range(1, DUNGEON_Y - 2);
     } while (charpair(dest) || mappair(dest) < ter_floor);
   }
 
-  if (charpair(dest) && charpair(dest) != d->PC) {
+  if (charpair(dest) && charpair(dest) != d->PC)
+  {
     io_queue_message("Teleport failed.  Destination occupied.");
-  } else {  
+  }
+  else
+  {
     d->character_map[d->PC->position[dim_y]][d->PC->position[dim_x]] = NULL;
     d->character_map[dest[dim_y]][dest[dim_x]] = d->PC;
 
@@ -527,36 +877,36 @@ uint32_t io_teleport_pc(dungeon *d)
 
 /* Adjectives to describe our monsters */
 static const char *adjectives[] = {
-  "A menacing ",
-  "A threatening ",
-  "A horrifying ",
-  "An intimidating ",
-  "An aggressive ",
-  "A frightening ",
-  "A terrifying ",
-  "A terrorizing ",
-  "An alarming ",
-  "A dangerous ",
-  "A glowering ",
-  "A glaring ",
-  "A scowling ",
-  "A chilling ",
-  "A scary ",
-  "A creepy ",
-  "An eerie ",
-  "A spooky ",
-  "A slobbering ",
-  "A drooling ",
-  "A horrendous ",
-  "An unnerving ",
-  "A cute little ",  /* Even though they're trying to kill you, */
-  "A teeny-weenie ", /* they can still be cute!                 */
-  "A fuzzy ",
-  "A fluffy white ",
-  "A kawaii ",       /* For our otaku */
-  "Hao ke ai de ",   /* And for our Chinese */
-  "Eine liebliche "  /* For our Deutch */
-  /* And there's one special case (see below) */
+    "A menacing ",
+    "A threatening ",
+    "A horrifying ",
+    "An intimidating ",
+    "An aggressive ",
+    "A frightening ",
+    "A terrifying ",
+    "A terrorizing ",
+    "An alarming ",
+    "A dangerous ",
+    "A glowering ",
+    "A glaring ",
+    "A scowling ",
+    "A chilling ",
+    "A scary ",
+    "A creepy ",
+    "An eerie ",
+    "A spooky ",
+    "A slobbering ",
+    "A drooling ",
+    "A horrendous ",
+    "An unnerving ",
+    "A cute little ",  /* Even though they're trying to kill you, */
+    "A teeny-weenie ", /* they can still be cute!                 */
+    "A fuzzy ",
+    "A fluffy white ",
+    "A kawaii ",      /* For our otaku */
+    "Hao ke ai de ",  /* And for our Chinese */
+    "Eine liebliche " /* For our Deutch */
+                      /* And there's one special case (see below) */
 };
 
 static void io_scroll_monster_list(char (*s)[40], uint32_t count)
@@ -566,25 +916,29 @@ static void io_scroll_monster_list(char (*s)[40], uint32_t count)
 
   offset = 0;
 
-  while (1) {
-    for (i = 0; i < 13; i++) {
+  while (1)
+  {
+    for (i = 0; i < 13; i++)
+    {
       mvprintw(i + 6, 19, " %-40s ", s[i + offset]);
     }
-    switch (getch()) {
+    switch (getch())
+    {
     case KEY_UP:
-      if (offset) {
+      if (offset)
+      {
         offset--;
       }
       break;
     case KEY_DOWN:
-      if (offset < (count - 13)) {
+      if (offset < (count - 13))
+      {
         offset++;
       }
       break;
     case 27:
       return;
     }
-
   }
 }
 
@@ -593,9 +947,9 @@ static void io_list_monsters_display(dungeon *d,
                                      uint32_t count)
 {
   uint32_t i;
-  char (*s)[40]; /* pointer to array of 40 char */
+  char(*s)[40]; /* pointer to array of 40 char */
 
-  s = (char (*)[40]) malloc(count * sizeof (*s));
+  s = (char(*)[40])malloc(count * sizeof(*s));
 
   mvprintw(3, 19, " %-40s ", "");
   /* Borrow the first element of our array for this string: */
@@ -603,31 +957,32 @@ static void io_list_monsters_display(dungeon *d,
   mvprintw(4, 19, " %-40s ", s);
   mvprintw(5, 19, " %-40s ", "");
 
-  for (i = 0; i < count; i++) {
+  for (i = 0; i < count; i++)
+  {
     snprintf(s[i], 40, "%16s%c: %2d %s by %2d %s",
-             (c[i]->symbol == 'd' ? "A tenacious " :
-              adjectives[rand() % (sizeof (adjectives) /
-                                   sizeof (adjectives[0]))]),
+             (c[i]->symbol == 'd' ? "A tenacious " : adjectives[rand() % (sizeof(adjectives) / sizeof(adjectives[0]))]),
              c[i]->symbol,
              abs(c[i]->position[dim_y] - d->PC->position[dim_y]),
-             ((c[i]->position[dim_y] - d->PC->position[dim_y]) <= 0 ?
-              "North" : "South"),
+             ((c[i]->position[dim_y] - d->PC->position[dim_y]) <= 0 ? "North" : "South"),
              abs(c[i]->position[dim_x] - d->PC->position[dim_x]),
-             ((c[i]->position[dim_x] - d->PC->position[dim_x]) <= 0 ?
-              "West" : "East"));
-    if (count <= 13) {
+             ((c[i]->position[dim_x] - d->PC->position[dim_x]) <= 0 ? "West" : "East"));
+    if (count <= 13)
+    {
       /* Handle the non-scrolling case right here. *
        * Scrolling in another function.            */
       mvprintw(i + 6, 19, " %-40s ", s[i]);
     }
   }
 
-  if (count <= 13) {
+  if (count <= 13)
+  {
     mvprintw(count + 6, 19, " %-40s ", "");
     mvprintw(count + 7, 19, " %-40s ", "Hit escape to continue.");
     while (getch() != 27 /* escape */)
       ;
-  } else {
+  }
+  else
+  {
     mvprintw(19, 19, " %-40s ", "");
     mvprintw(20, 19, " %-40s ",
              "Arrows to scroll, escape to continue.");
@@ -642,14 +997,17 @@ static void io_list_monsters(dungeon *d)
   character **c;
   uint32_t x, y, count;
 
-  c = (character **) malloc(d->num_monsters * sizeof (*c));
+  c = (character **)malloc(d->num_monsters * sizeof(*c));
 
   /* Get a linear list of monsters */
-  for (count = 0, y = 1; y < DUNGEON_Y - 1; y++) {
-    for (x = 1; x < DUNGEON_X - 1; x++) {
+  for (count = 0, y = 1; y < DUNGEON_Y - 1; y++)
+  {
+    for (x = 1; x < DUNGEON_X - 1; x++)
+    {
       if (d->character_map[y][x] && d->character_map[y][x] != d->PC &&
           can_see(d, character_get_pos(d->PC),
-                  character_get_pos(d->character_map[y][x]), 1, 0)) {
+                  character_get_pos(d->character_map[y][x]), 1, 0))
+      {
         c[count++] = d->character_map[y][x];
       }
     }
@@ -657,7 +1015,7 @@ static void io_list_monsters(dungeon *d)
 
   /* Sort it by distance from PC */
   thedungeon = d;
-  qsort(c, count, sizeof (*c), compare_monster_distance);
+  qsort(c, count, sizeof(*c), compare_monster_distance);
 
   /* Display it */
   io_list_monsters_display(d, c, count);
@@ -672,8 +1030,10 @@ void io_handle_input(dungeon *d)
   uint32_t fail_code;
   int key;
 
-  do {
-    switch (key = getch()) {
+  do
+  {
+    switch (key = getch())
+    {
     case '7':
     case 'y':
     case KEY_HOME:
@@ -763,7 +1123,7 @@ void io_handle_input(dungeon *d)
       io_display_no_fog(d);
       fail_code = 1;
       break;
-     case 'm':
+    case 'm':
       io_list_monsters(d);
       fail_code = 1;
       break;